+++ frontend/src/hooks/useOutsideClick.ts
@@ -1,0 +1,20 @@
+import { useCallback, useEffect } from 'react';
+
+export const useOutsideClick = (
+  closeModal: () => void,
+  ref: React.RefObject<HTMLDivElement>
+) => {
+  const checkIfClickedOutside = useCallback(
+    ({ target }: MouseEvent) => {
+      if (!ref.current?.contains(target as Node)) {
+        closeModal();
+      }
+    },
+    [ref, closeModal]
+  );
+
+  useEffect(() => {
+    window.addEventListener('mouseup', checkIfClickedOutside);
+    return () => window.removeEventListener('mouseup', checkIfClickedOutside);
+  }, [checkIfClickedOutside]);
+};
