+++ frontend/src/hooks/useRemoveScrollWhenModalOpen.ts
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+
+export const useRemoveScrollWhenModalOpen = (isOpen: boolean) => {
+  useEffect(() => {
+    if (isOpen) {
+      document.body.style.overflow = 'hidden';
+    } else {
+      document.body.style.overflow = 'unset';
+    }
+  }, [isOpen]);
+};
